Role key type is now Integer instead Long

+++ src/main/java/ru/eninja/dao/RoleDao.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import ru.eninja.domain.Role;
 
 
-public interface RoleDao extends CrudRepository<Role, Long> {
+public interface RoleDao extends CrudRepository<Role, Integer> {
 }
